Refactor map logic into its own file.
