+++ map.h
@@ -1,0 +1,8 @@
+#ifndef MAP_H
+#define MAP_H
+
+void init_map(void *level_data);
+void draw_map_screen();
+void draw_map();
+
+#endif /* MAP_H */
